src/agents/visual_agent/pyproject.toml
[project]
name = "visual-agent"
version = "0.1.0"
description = "Visual agent"
authors = [{ name = "brunomaso1", email = "brunomaso1@hotmail.com" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">3.10,<4.0"
dependencies = [
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "openai (>=1.109.1,<2.0.0)",
    "a2a-sdk[http-server] (>=0.3.7,<0.4.0)",
    "click (>=8.3.0,<9.0.0)",
    "uvicorn (>=0.37.0,<0.38.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "opencv-python (>=4.12.0.88,<5.0.0.0)",
]

[tool.poetry]
packages = [{ include = "src" }]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
